USB: android: Avoid re-enabling the same configuration
ADB reader is unblocked and -EIO is returned to user space
upon cable disconnect or receiving reset interrupt in configured
state.  ADB daemon re-open the port which is causing re-enabling
the same configuration.  The pull-up is disabled till all the
function drivers are re-enabled.  The USB flakiness is observed
with some hosts.

Delay adb close event to android by 1 sec. If ADB opens the port
with in 1sec, cancel notifying this event.

Change-Id: Ibd179a39019bd6e7e472df4430cba5845fac3cf2
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>